+++ dslink-v2-websocket/src/main/java/org/iot/dsa/dslink/websocket/StandaloneTransportFactory.java
@@ -1,10 +1,7 @@
 package org.iot.dsa.dslink.websocket;
 
-import com.acuity.iot.dsa.dslink.transport.DSBinaryTransport;
-import com.acuity.iot.dsa.dslink.transport.DSTextTransport;
 import com.acuity.iot.dsa.dslink.transport.DSTransport.Factory;
 import org.iot.dsa.dslink.DSLinkConnection;
-import org.iot.dsa.logging.DSLogger;
 
 /**
  * Websocket client implementation of DSTransport.Factory based on Tyrus, the reference
@@ -12,16 +9,11 @@ import org.iot.dsa.logging.DSLogger;
  *
  * @author Aaron Hansen
  */
-public class StandaloneTransportFactory extends DSLogger implements Factory {
+public class StandaloneTransportFactory implements Factory {
 
     @Override
-    public DSBinaryTransport makeBinaryTransport(DSLinkConnection conn) {
-        return new WsBinaryTransport();
-    }
-
-    @Override
-    public DSTextTransport makeTextTransport(DSLinkConnection conn) {
-        return new WsTextTransport();
+    public TyrusTransport makeTransport(DSLinkConnection conn) {
+        return new TyrusTransport();
     }
 
 }
