+++ dslink-v2/src/main/java/com/acuity/iot/dsa/dslink/io/msgpack/MsgpackWriter.java
@@ -1,7 +1,7 @@
 package com.acuity.iot.dsa.dslink.io.msgpack;
 
 import com.acuity.iot.dsa.dslink.io.DSByteBuffer;
-import com.acuity.iot.dsa.dslink.transport.DSBinaryTransport;
+import com.acuity.iot.dsa.dslink.transport.DSTransport;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.ByteBuffer;
@@ -84,7 +84,7 @@ public class MsgpackWriter extends AbstractWriter implements MsgpackConstants {
     /**
      * Writes the internal buffer to the parameter.  The internal buffer will be cleared.
      */
-    public void writeTo(DSBinaryTransport out) {
+    public void writeTo(DSTransport out) {
         byteBuffer.sendTo(out, (frame == null));
     }
