+++ dslink-v2/src/main/java/com/acuity/iot/dsa/dslink/protocol/v2/requester/DS2OutboundInvokeStub.java
@@ -8,7 +8,6 @@ import com.acuity.iot.dsa.dslink.protocol.v2.DS2MessageReader;
 import com.acuity.iot.dsa.dslink.protocol.v2.DS2MessageWriter;
 import com.acuity.iot.dsa.dslink.protocol.v2.MessageConstants;
 import com.acuity.iot.dsa.dslink.protocol.v2.MultipartWriter;
-import com.acuity.iot.dsa.dslink.transport.DSBinaryTransport;
 import org.iot.dsa.dslink.requester.OutboundInvokeHandler;
 import org.iot.dsa.node.DSMap;
 
@@ -53,7 +52,7 @@ public class DS2OutboundInvokeStub extends DSOutboundInvokeStub
             multipart.update(out, ack);
             getRequester().sendRequest(this);
         } else {
-            out.write((DSBinaryTransport) getRequester().getTransport());
+            out.write(getRequester().getTransport());
         }
     }
