+++ dslink-v2/src/main/java/com/acuity/iot/dsa/dslink/protocol/v2/responder/DS2InboundList.java
@@ -9,7 +9,6 @@ import com.acuity.iot.dsa.dslink.protocol.responder.DSInboundList;
 import com.acuity.iot.dsa.dslink.protocol.v2.DS2MessageWriter;
 import com.acuity.iot.dsa.dslink.protocol.v2.MessageConstants;
 import com.acuity.iot.dsa.dslink.protocol.v2.MultipartWriter;
-import com.acuity.iot.dsa.dslink.transport.DSBinaryTransport;
 import org.iot.dsa.node.DSElement;
 import org.iot.dsa.node.DSPath;
 
@@ -44,7 +43,7 @@ class DS2InboundList extends DSInboundList implements MessageConstants {
             getResponder().sendResponse(this);
         } else {
             DSBrokerConnection up = (DSBrokerConnection) getResponder().getConnection();
-            out.write((DSBinaryTransport) up.getTransport());
+            out.write(up.getTransport());
         }
     }
