+++ dslink-v2/src/main/java/com/acuity/iot/dsa/dslink/test/V1TestLink.java
@@ -4,6 +4,7 @@ import com.acuity.iot.dsa.dslink.protocol.DSRootLink;
 import com.acuity.iot.dsa.dslink.protocol.v1.DS1ConnectionInit;
 import com.acuity.iot.dsa.dslink.protocol.v1.DS1LinkConnection;
 import com.acuity.iot.dsa.dslink.transport.DSTransport;
+import java.util.logging.Level;
 import org.iot.dsa.dslink.DSLinkOptions;
 import org.iot.dsa.dslink.DSMainNode;
 import org.iot.dsa.node.DSMap;
@@ -40,17 +41,8 @@ public class V1TestLink extends DSRootLink {
             return new DS1ConnectionInit();
         }
 
-        /**
-         * Looks at the connection initialization response to determine the type of transport then
-         * instantiates the correct type fom the config.
-         */
         protected DSTransport makeTransport(DS1ConnectionInit init) {
-            TestTransport transport = new TestTransport();
-            transport.setConnection(this);
-            transport.setReadTimeout(getLink().getOptions().getConfig(
-                    DSLinkOptions.CFG_READ_TIMEOUT, 60000));
-            setTransport(transport);
-            return transport;
+            return new TestTransport().setText(true);
         }
 
     }
