+++ dslink-v2/src/main/java/org/iot/dsa/node/DSNode.java
@@ -695,6 +695,7 @@ public class DSNode extends DSLogger implements DSIObject, Iterable<DSInfo> {
      * @see org.iot.dsa.node.action.DSAction#invoke(DSInfo, ActionInvocation)
      */
     public ActionResult invoke(DSInfo action, DSInfo target, ActionInvocation request) {
+        trace(trace() ? String.format("action=%s, target=%s, params=%s", action, target, request.getParameters()) : null);
         return action.getAction().invoke(target, request);
     }
 
@@ -1214,6 +1215,7 @@ public class DSNode extends DSLogger implements DSIObject, Iterable<DSInfo> {
         if (event == null) {
             throw new NullPointerException("Null event");
         }
+        trace(trace() ? String.format("event=%s, child=%s, data=%s", event.getEventId(), child, data) : null);
         Subscription sub = subscription;
         while (sub != null) {
             try {
